Pull out the rate limiting into a generic utility class
